+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes qsf-integration project will be documented in this file.
 ### Added
 - make it possible to set a custom searchQuery object in the elastic query transformer
 - implement tracking for range values
+- implement date range queries for qsfql and elasticsearch
 
 
 ### Changed